add role middleware

+++ routes/V1/admin_routes.php
@@ -3,33 +3,33 @@
 use Illuminate\Support\Facades\Route;
 
 Route::prefix('admin')->namespace('App\Http\Controllers\API\V1\Admin')
-    ->middleware(['clerk.auth:api', 'role:admin'])
+    ->middleware(['clerk.auth:api', 'role:admin,superadmin,staff'])
     ->group(function () {
         // Admin Routes
         Route::get('dashboard', 'DashboardController@index');
         Route::get('clerk/test', fn() => 'Clerk Middleware Check Admin | ' . auth()->user()->clerk_id);
         Route::apiResource('rooms', 'RoomController');
 
-        Route::apiResource('users', 'UserController');
+        Route::apiResource('users', 'UserController')->middleware('role:admin,superadmin');
 
-        Route::apiResource('amenities', 'AmenityController');
-        Route::patch('amenities/{id}/update-status', 'AmenityController@updateStatus');
+        Route::apiResource('amenities', 'AmenityController')->middleware('role:admin,superadmin');
+        Route::patch('amenities/{id}/update-status', 'AmenityController@updateStatus')->middleware('role:admin,superadmin');
 
-        Route::apiResource('bookings', 'BookingController');
-        Route::post('bookings/{booking}/other-charges', 'BookingController@storeOtherCharge');
-        Route::patch('bookings/{booking}/reschedule', 'BookingController@reschedule');
-        Route::delete('bookings/{booking}/other-charges/{charge}', 'OtherChargeController@destroy');
+        Route::apiResource('bookings', 'BookingController')->middleware('role:admin,superadmin');
+        Route::post('bookings/{booking}/other-charges', 'BookingController@storeOtherCharge')->middleware('role:admin,superadmin');
+        Route::patch('bookings/{booking}/reschedule', 'BookingController@reschedule')->middleware('role:admin,superadmin');
+        Route::delete('bookings/{booking}/other-charges/{charge}', 'OtherChargeController@destroy')->middleware('role:admin,superadmin');
         // Route::apiResource('payments', 'PaymentController');
-        Route::post('payments/pay', 'PaymentController@pay');
-        Route::put('payments/{payment}', 'PaymentController@update');
+        Route::post('payments/pay', 'PaymentController@pay')->middleware('role:admin,superadmin');
+        Route::put('payments/{payment}', 'PaymentController@update')->middleware('role:admin,superadmin');
 
-        Route::patch('promos/bulk-update-status', 'PromoController@bulkUpdateStatus');
-        Route::patch('promos/{id}/update-status', 'PromoController@updateStatus');
-        Route::patch('promos/{id}/update-exclusive', 'PromoController@updateExclusive');
-        Route::apiResource('promos', 'PromoController');
+        Route::patch('promos/bulk-update-status', 'PromoController@bulkUpdateStatus')->middleware('role:admin,superadmin');
+        Route::patch('promos/{id}/update-status', 'PromoController@updateStatus')->middleware('role:admin,superadmin');
+        Route::patch('promos/{id}/update-exclusive', 'PromoController@updateExclusive')->middleware('role:admin,superadmin');
+        Route::apiResource('promos', 'PromoController')->middleware('role:admin,superadmin');
         
-        Route::apiResource('meal-prices', 'MealPriceController');
+        Route::apiResource('meal-prices', 'MealPriceController')->middleware('role:admin,superadmin');
 
-        Route::apiResource('images', 'ImageController')->only(['index', 'store', 'destroy']);
+        Route::apiResource('images', 'ImageController')->only(['index', 'store', 'destroy'])->middleware('role:admin,superadmin');
 
     });
